Move ef-tests into tests dir
