--- crates/evm/src/ef_tests/assert.rs
@@ -1,19 +1,0 @@
-//! Various assertion helpers.
-
-use std::fmt::Debug;
-
-use super::Error;
-
-/// A helper like `assert_eq!` that instead returns `Err(Error::Assertion)` on failure.
-pub fn assert_equal<T>(left: T, right: T, msg: &str) -> Result<(), Error>
-where
-    T: PartialEq + Debug,
-{
-    if left == right {
-        Ok(())
-    } else {
-        Err(Error::Assertion(format!(
-            "{msg}\n  left `{left:?}`,\n right `{right:?}`"
-        )))
-    }
-}
