+++ crates/evm/src/hooks.rs
@@ -80,8 +80,8 @@ where
         self.block_env.set(&new_pending_env, working_set);
         self.l1_fee_rate.set(&l1_fee_rate, working_set);
 
-        #[cfg(not(feature = "ef-tests"))]
         if !system_events.is_empty() {
+            // #[cfg(all(test, not(feature = "ef-tests")))]
             self.execute_system_events(system_events, working_set);
         }
