+++ crates/evm/src/lib.rs
@@ -6,8 +6,6 @@ mod genesis;
 mod hooks;
 mod provider_functions;
 mod rpc_helpers;
-#[cfg(any(test, feature = "ef-tests"))]
-mod test_utils;
 pub use call::*;
 pub use error::rpc::*;
 pub use evm::*;
@@ -22,8 +20,6 @@ pub use signer::DevSigner;
 #[cfg(feature = "smart_contracts")]
 pub mod smart_contracts;
 pub use system_events::SYSTEM_SIGNER;
-// #[cfg(feature = "ef-tests")]
-pub mod ef_tests;
 
 #[cfg(test)]
 mod tests;
