+++ crates/evm/src/tests/ef_tests/mod.rs
@@ -1,7 +1,25 @@
+//! Abstractions and runners for EF tests.
+
+#![allow(invalid_doc_attributes)]
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
+    html_favicon_url = "https://avatars0.githubusercontent.com/u/97369466?s=256",
+    issue_tracker_base_url = "https://github.com/paradigmxyz/reth/issues/"
+)]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg(feature = "ef-tests")]
 
-use citrea_evm::ef_tests::cases::blockchain_test::BlockchainTests;
-use citrea_evm::ef_tests::suite::Suite;
+pub mod case;
+pub mod result;
+pub mod suite;
+
+pub mod cases;
+pub mod models;
+
+pub use case::Case;
+use cases::blockchain_test::BlockchainTests;
+pub use result::Error;
+pub use suite::Suite;
 
 macro_rules! general_state_test {
     ($test_name:ident, $dir:ident) => {
