+++ crates/evm/src/tests/mod.rs
@@ -1,8 +1,10 @@
 mod call_tests;
 mod cfg_tests;
+mod ef_tests;
 mod genesis_tests;
 mod hooks_tests;
 mod queries;
 mod sys_tx_tests;
 pub(crate) mod test_signer;
 mod tx_tests;
+mod utils;
